+++ iceoryx_examples/complexdata/iox_publisher_vector.cpp
@@ -51,7 +51,7 @@ int main()
             })
             .or_else([](auto& error) {
                 // do something with error
-                std::cerr << "Unable to loan sample, error code: " << static_cast<uint64_t>(error) << std::endl;
+                std::cerr << "Unable to loan sample, error code: " << error << std::endl;
             });
         ++ct;
