+++ iceoryx_examples/user_header/publisher_cxx_api.cpp
@@ -61,8 +61,7 @@ int main()
             })
             .or_else([&](auto& error) {
                 //! [loan failed]
-                std::cout << APP_NAME << " could not loan sample! Error code: " << static_cast<uint64_t>(error)
-                          << std::endl;
+                std::cout << APP_NAME << " could not loan sample! Error code: " << error << std::endl;
                 //! [loan failed]
             });
         //! [loan sample]
